🧑‍💻 move editor settings method next to the `new` method
